fix: increase event listeners on the socket
There can be as many as N+X listeners attached to the underlying socket.
Where N is the max number of channels, and X is a small variable. I
don't want to make little adjustments to the maxListeners so I'll just
make it unlimited and take the small risk of a leak for now.